rtl8723bu: indicate disconnection events when disconnecting
For some reason, the driver has logic to suppress raising disconnection
events when in these codepaths.

This causes mac80211 to believe the device is still connected. When the
user attempts to switch from one AP to another, the 2nd connection
will fail because of this confusion.

Follow the approach taken in newer Realtek drivers for other hardware,
where these events are only suppressed in the connection codepath,
and issued as normal during disconnection.